fix(EI-314): /api/v1/profile and /overflow 500 for any account that resized the board (#106)
Every Raycast command was failing with a 500. Captured from production logs
rather than guessed:

  "path": ["splitRatio"], "expected": "int",
  "message": "Invalid input: expected int, received number"

`clampSplit` (`use-split-resize.ts`) returns `(topPx / totalPx) * 100` and
never rounds, so the board has written a FRACTION into `splitRatio` since the
resize seam shipped. `schema.ts` declared `.int()`. The two have disagreed
the whole time.

Nothing caught it because nothing validated. The client's `mutate()` writes
to Dexie without running the schema, `sanitizePatch` is a column whitelist
rather than a validator, and SQLite's `integer` column type is advisory — it
stores a real without complaint. `/api/v1/profile` (EI-297) was simply the
first code to PARSE a settings row, and it 500'd for every account that had
ever dragged the divider. MCP's `get_profile`/`get_overflow` have been broken
the same way since A6, silently, because nothing called them.

Rounding would have been the wrong fix: at 1% granularity a drag on a 900px
board jumps nine pixels at a time. The schema was describing data that never
existed, so the schema is what changes.

The second half matters more than the first. `settingsOrDefault` used
`parse()`, so ONE bad field took down every caller — and `/api/v1/profile`
does not even EXPOSE `splitRatio`, since it is device-local. A read that
wants nine fields failed on a tenth it deliberately withholds.

A settings row is a wide bag of mostly-cosmetic preferences, and over the
life of a synced product the odds of some field drifting from its declared
type are high. So it now strips the fields Zod rejects, keeps everything that
parses, and lets the schema's own defaults fill the gaps — bounded, since
each pass removes at least one key. A reader that wants a timezone no longer
cares that a pane divider is half a percent off.

Also caught in passing: my own verification could not have found this. The 50
live assertions in `scripts/v1-smoke/` ran against a freshly created account
with NO settings row, which takes the `?? { ownerId, updatedAt }` path and
never parses real data. Worth remembering when a smoke harness looks green.

2,683 tests pass; openapi regenerated for the widened type.

+++ src/lib/schema.ts
@@ -648,7 +648,23 @@ export const settingsSchema = z.object({
    * lib/sync/wire.ts): a percentage transfers between a laptop and a wide
    * monitor on the same account where a pixel width would not.
    */
-  splitRatio: z.number().int().min(SPLIT_MIN_PERCENT).max(SPLIT_MAX_PERCENT).nullable().default(null),
+  /**
+   * NOT `.int()`. `clampSplit` (`use-split-resize.ts`) returns
+   * `(topPx / totalPx) * 100` and never rounds, so the board has always
+   * written a fraction here — the integer constraint described data that has
+   * never existed.
+   *
+   * Nothing caught it because nothing validated: the client's `mutate()`
+   * writes to Dexie without running this schema, `sanitizePatch` is a column
+   * whitelist rather than a validator, and SQLite's `integer` column type is
+   * advisory — it stores a real quite happily. The first code to actually
+   * PARSE a settings row was `/api/v1/profile` (EI-297), which then 500'd for
+   * every account that had ever dragged the seam (EI-314).
+   *
+   * Rounding instead would be the wrong fix: at 1% granularity a drag on a
+   * 900px board would jump nine pixels at a time.
+   */
+  splitRatio: z.number().min(SPLIT_MIN_PERCENT).max(SPLIT_MAX_PERCENT).nullable().default(null),
   /**
    * Which half, if either, is collapsed. An enum rather than two booleans so
    * "both collapsed at once" is unrepresentable rather than merely avoided.
